Reorganize the project structure.
