+++ examples/package.json
@@ -1,0 +1,5 @@
+{
+  "dependencies": {
+    "delta-sharing": "^0.0.1"
+  }
+}
